+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CORS Policy regarding REST server implementation
 - Fixed some bugs in multimodal usecase
 - Fixed File datatype
+- Fixed a bug in artifact store to skip duplicate artifacts
 
 
 ## [0.1.1](https://github.com/SuperDuperDB/superduperdb/compare/0.0.20...0.1.0])    (2023-Feb-09)